refactor: switch POST to PATCH and remove id from committime api routes
